Add option to prompt for status instead of extract
The part of visiting Discord is kind of buggy not because of AppleScript
but rather because Discord sometimes logs me out and/or resets my status
when I do log in, causing an error. We make the extraction step now
opt-in, defaulting to dialog boxes that prompt for the text and the
emoji. The text part defaults to the system clipboard, so the new flow
looks something like:

    1. Copy status from Discord client (like I used to do anyway).
    2. Launch application.
    3. Hit ENTER to just accept the status text pre-populated from
       clipboard.
    4. Add emoji if any and hit ENTER.
    5. Upload proceeds like normal.